requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
llama-index>=0.10.0
llama-index-vector-stores-milvus>=0.1.0
llama-index-embeddings-dashscope==0.4.1
llama-index-llms-dashscope==0.5.2
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
pymilvus>=2.4.0
sentence-transformers>=2.2.0
FlagEmbedding>=1.2.0
tiktoken>=0.5.0
unstructured>=0.10.0
pdfplumber>=0.10.0
python-docx>=0.8.11
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
jieba>=0.42.1
rank-bm25>=0.2.2
ragas>=0.1.0
pydantic>=2.0
python-dotenv>=1.0.0
rich>=13.0.0